utilities/wavToFlac/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
mutagen==1.46.0
oauthlib==3.2.2
python-dateutil==2.8.2
python3-discogs-client==2.6
requests==2.28.2
six==1.16.0
urllib3==1.26.14